Реорганизация door.*
