+++ server/CMakeLists.txt
@@ -12,11 +12,7 @@ list(APPEND SVDLL_SOURCES
 	"client.cpp"
 	"combat.cpp"
 	"dll_int.cpp"
-	"doors.cpp"
 	"effects.cpp"
-	"explode.cpp"
-	"func_break.cpp"
-	"func_monitor.cpp"
 	"func_tank.cpp"
 	"game.cpp"
 	"globals.cpp"
